\name{colMin}
\alias{colMin}
\alias{rowMin}
\title{Form Row and Column Min and Max}
\usage{
colMin(x, na.rm = FALSE)

rowMin(x, na.rm = FALSE)
}
\arguments{
  \item{x}{an array of two or more dimensions, containing
  numeric values}

  \item{na.rm}{logical. Should missing values (including
  NaN) be omitted from the calculations?}
}
\description{
Form row and column min and max for numeric arrays.
}

