% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{drtools-package}
\alias{drtools}
\alias{drtools-package}
\title{drtools: dr's tools for R}
\description{
A compilation of different tools for R
}
\details{
\tabular{ll}{
  Package: \tab drtools \cr
  Version: \tab 0.0.1 \cr
  License: \tab GPL-3 \cr
}
}
\author{
Daniel Rodriguez Perez \email{daniel.rodriguez.perez@gmail.com}
}

